Fixes #22538 - 'repo' should mention filesystem too
The "Repo" setting can be used to export/import from a filesystem,
ssh, etc..
The documentation and the repository README explain that if it
points to '/tmp/' (just an example) it will export templates to
that destination. I tested it and it works well.
However, on Administer > Settings, the text explicitly says "Git
repo", which may confuse an user into thinking they can only use a
Git repo as the target.