Fixes #37986, #37982 - Fix UI Error when importing templates
Importing templates through UI causes an error and the page with import
results does not load.

Steps to Reproduce:

1. Have foreman instance with the templates plugin on nightly (error
originates in one of the newest commits so it is not on stream yet).

2. Navigate to Hosts > Templates > Sync Templates

3. Import templates from
https://github.com/theforeman/community-templates.git

Actual behavior:

The page with results does not load and two popups are displayed. One
saying that templates have been imported succesfully and another saying
that there has been an error.

Expected behavior:

Page with results loads normally and displays templates correctly.