Fixes #38195 - Hide http proxy setting when syncing through ssh
Currently, the HTTP proxy policy setting in ui is displayed at all
times. However it only goes into effect when syncing through HTTP or
HTTPS. SSH sync does not use the proxy, therefore the user shoudn't be
able to see the proxy policy if a protocol different from HTTP(S) is
specified in the repo url.